Create CTEs
A common table expression (CTE) is a temporary table you can reference in the main query that follows it. It’s a lot like a subquery, but with one crucial difference: you write the query to create your CTE and assign it a name at the start of the main query. To access the CTE in the main query that follows, you use its name just as you would any other table in the database.